inference/tts/pyproject.toml
[project]
name = "whats-tts"
version = "0.1.0"
description = "TTS (text-to-speech) gRPC service"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.67.0,<2.0",
    "grpcio-tools>=1.67.0,<2.0",
    "protobuf>=5.28.0,<6.0",
    "numpy>=1.26.0,<2.0",
    # TTS engine TBD — add dependency when engine is chosen in Phase 2
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
